Namespace bot API controllers under API::Bots with nested resources
