+++ app/controllers/api/bots/rooms_controller.rb
@@ -1,0 +1,62 @@
+class API::Bots::RoomsController < API::Bots::BaseController
+  before_action :set_room, only: %i[update destroy]
+  before_action :require_creator, only: %i[update destroy]
+
+  def index
+    @rooms = if params[:joinable].present?
+      Rooms::Open.browsable_by(Current.user)
+    else
+      Current.user.rooms.without_threads
+    end
+  end
+
+  def create
+    type = params[:type]&.downcase
+
+    unless type.in?(%w[open closed])
+      return render json: { error: "Type must be 'open' or 'closed'", code: "validation_failed" }, status: :unprocessable_entity
+    end
+
+    if params[:name].blank?
+      return render json: { error: "Name is required", code: "validation_failed" }, status: :unprocessable_entity
+    end
+
+    room_class = type == "open" ? Rooms::Open : Rooms::Closed
+    @room = room_class.create_for({ name: params[:name] }, users: Current.user)
+
+    render :create, status: :created
+  end
+
+  def update
+    if params[:name].blank?
+      return render json: { error: "Name is required", code: "validation_failed" }, status: :unprocessable_entity
+    end
+
+    old_name = @room.name
+    @room.update!(name: params[:name])
+    @room.announce_rename(old_name, actor: Current.user) if @room.name != old_name
+    RoomUpdateBroadcastJob.perform_later(@room)
+
+    render :update
+  end
+
+  def destroy
+    @room.deactivate
+    broadcast_remove_room
+
+    head :no_content
+  rescue Room::CannotDeleteOriginalError
+    render json: { error: "The original room can't be deleted", code: "validation_failed" }, status: :unprocessable_entity
+  end
+
+  private
+    def set_room
+      @room = Current.user.rooms.find(params[:id])
+    end
+
+    def broadcast_remove_room
+      Sidebar::SIDEBAR_SECTIONS.each do |list_name|
+        broadcast_remove_to Current.account, :rooms, target: [ @room, helpers.dom_prefix(list_name, :list_node) ]
+      end
+    end
+end
