+++ app/models/room.rb
@@ -76,10 +76,6 @@ class Room < ApplicationRecord
     end
   end
 
-  def as_bot_json(url_helper:)
-    { id: id, name: name, type: type.demodulize, messages_url: url_helper.call(self) }
-  end
-
   def receive(message)
     unread_memberships(message)
     push_later(message)
