+++ app/views/api/bots/rooms/_room.json.jbuilder
@@ -1,0 +1,4 @@
+json.id   room.id
+json.name room.name
+json.type room.type.demodulize
+json.messages_url api_bots_room_messages_url(room)
