+++ app/views/api/bots/rooms/create.json.jbuilder
@@ -1,0 +1,1 @@
+json.partial! "api/bots/rooms/room", room: @room
