+++ app/views/api/bots/rooms/index.json.jbuilder
@@ -1,0 +1,1 @@
+json.array! @rooms, partial: "api/bots/rooms/room", as: :room
